Начал вводить копейки в денежную систему. Исправил валидацию и добавил сообщения об ошибках

+++ sql/events.sql
@@ -2,7 +2,7 @@ CREATE DATABASE orsys_events CHARACTER SET utf8 COLLATE utf8_general_ci;
 USE orsys_events;
 
 CREATE TABLE events (
-  `id` INT NOT NULL AUTO_INCREMENT,
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
   `date` TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
   `event` TEXT NOT NULL,
   INDEX(`date`),
