+++ sql/orders.sql
@@ -2,7 +2,7 @@ CREATE DATABASE orsys_orders CHARACTER SET utf8 COLLATE utf8_unicode_ci;
 USE orsys_orders;
 
 CREATE TABLE orders (
-  `id` INT NOT NULL AUTO_INCREMENT,
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
   `client` INT NOT NULL,
   `description` TEXT NOT NULL,
   `cost` INT NOT NULL,
